Normalize file structure to match dev to make future merging easier
